requirements.txt
aiogram==3.2.0
sqlalchemy==2.0.19
python-dotenv==1.0.0
